refactor registry to lookup serverice
